intro_drf/requirements.txt
Django==4.2.7
psycopg2==2.9.9
django-environ==0.11.2
django-redis==5.4.0
djangorestframework==3.14.0